requirements.txt
asgiref
autopep8
Django==3.2.16
pycodestyle
pytz
sqlparse
Unipath
dj-database-url
python-decouple
gunicorn
whitenoise
schedule
honcho
firebase==3.0.1
firebase-admin==5.2.0
pandas==1.4.1
mysql-connector-python==8.0.32
numpy 
scikit-learn
matplotlib
openpyxl